Add instance JSON-RPC execution path
Introduce JsonRpcRequestProcessor and UnityCliLoopExecutionRouter so JSON-RPC execution can be injected without changing the existing static entrypoints yet. The compatibility wrappers read ApplicationRegistrar.Service at call time to preserve the current test swap seam.
