+++ Packages/src/Editor/Infrastructure/Api/GetToolDetailsBridgeCommand.cs
@@ -4,7 +4,6 @@ using Newtonsoft.Json.Linq;
 using io.github.hatayama.UnityCliLoop.Application;
 using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
-using ApplicationRegistrar = io.github.hatayama.UnityCliLoop.Application.UnityCliLoopToolRegistrar;
 
 namespace io.github.hatayama.UnityCliLoop.Infrastructure
 {
@@ -16,11 +15,13 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
         private const string IncludeDevelopmentOnlyPropertyName = "IncludeDevelopmentOnly";
         private const string IncludeDevelopmentOnlyCamelCasePropertyName = "includeDevelopmentOnly";
 
-        public static GetToolDetailsResponse Execute(JToken paramsToken)
+        public static GetToolDetailsResponse Execute(
+            JToken paramsToken,
+            UnityCliLoopToolRegistrarService toolRegistrarService)
         {
             bool includeDevelopmentOnly = ReadIncludeDevelopmentOnly(paramsToken);
 
-            UnityCliLoopToolRegistry registry = ApplicationRegistrar.GetRegistry();
+            UnityCliLoopToolRegistry registry = toolRegistrarService.GetRegistry();
             ToolInfo[] allTools = registry.GetRegisteredTools();
 
             ToolInfo[] filteredTools = allTools;
