+++ Packages/src/Editor/Infrastructure/Api/UnityApiHandler.cs
@@ -3,52 +3,24 @@ using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
 
 using io.github.hatayama.UnityCliLoop.Application;
-using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.ToolContracts;
 using ApplicationRegistrar = io.github.hatayama.UnityCliLoop.Application.UnityCliLoopToolRegistrar;
 
 namespace io.github.hatayama.UnityCliLoop.Infrastructure
 {
-
     /// <summary>
-    /// Routes JSON-RPC execution requests received by JsonRpcProcessor to either the
-    /// internal bridge command router or the registered tools in the Application layer.
-    /// Terminology for "tool" vs "internal bridge command" is defined in docs/glossary.md.
+    /// Compatibility entrypoint for execution callers that have not received UnityCliLoopExecutionRouter through DI yet.
     /// </summary>
     public static class UnityApiHandler
     {
-        /// <summary>
-        /// Get command registry
-        /// Use this registry when adding new commands
-        /// </summary>
-        public static UnityCliLoopToolRegistry CommandRegistry => ApplicationRegistrar.GetRegistry();
-
-        /// <summary>
-        /// Generic command execution method
-        /// Uses new command-based structure
-        /// </summary>
-        /// <param name="commandName">Command name</param>
-        /// <param name="paramsToken">Parameters</param>
-        /// <returns>Execution result</returns>
-        public static async Task<UnityCliLoopToolResponse> ExecuteCommandAsync(
+        public static Task<UnityCliLoopToolResponse> ExecuteCommandAsync(
             string commandName,
             JToken paramsToken,
             CancellationToken ct)
         {
-            UnityCliLoopToolResponse response;
-            if (InternalBridgeCommandRouter.IsInternalCommand(commandName))
-            {
-                await MainThreadSwitcher.SwitchToMainThread(ct);
-                ct.ThrowIfCancellationRequested();
-                response = InternalBridgeCommandRouter.Execute(commandName, paramsToken);
-                return response;
-            }
-
-            response = await ApplicationRegistrar.ExecuteToolAsync(
-                commandName,
-                paramsToken,
-                ct);
-            return response;
+            UnityCliLoopToolRegistrarService toolRegistrarService = ApplicationRegistrar.Service;
+            UnityCliLoopExecutionRouter executionRouter = new(toolRegistrarService);
+            return executionRouter.ExecuteAsync(commandName, paramsToken, ct);
         }
     }
 }
