+++ Packages/src/Editor/Infrastructure/Api/UnityApiHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 76ca8fbc661e34c98829974260660603
+guid: 5a0aa253bb5a497893c9623855b6ea6b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
